requirementsPip.txt
certifi==2020.6.20
chardet==3.0.4
facenet-pytorch==2.5.0
future==0.18.2
idna==2.10
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
numpy @ file:///C:/ci/numpy_and_numpy_base_1596215850360/work
olefile==0.46
Pillow @ file:///C:/ci/pillow_1594298230227/work
requests==2.24.0
six==1.15.0
torch==1.6.0
torchvision==0.7.0
urllib3==1.25.10
wincertstore==0.2
